update: refactorizacion
