+++ src/index.js
@@ -16,14 +16,14 @@ app.use('/greeting', greetingRouter);
 const authRouter = require('./routes/auth');
 app.use('/auth', authRouter);
 
-/* const travelRouter = require('./routes/travel');
-app.use('/travel', travelRouter); */
+const wanderlusRouter = require('./routes/wanderlust');
+app.use('/wanderlust', wanderlusRouter);
 
 const amadeusRouter = require('./routes/amadeus');
 app.use('/travel', amadeusRouter);
 
-const denarioRouter = require('./routes/denario');
-app.use('/denario', denarioRouter);
+/* const denarioRouter = require('./routes/denario');
+app.use('/denario', denarioRouter); */
 
 const usersRouter = require('./routes/users');
 app.use('/users', usersRouter);
